@@ -1,7 +1,7 @@
 class ProductsController < ApplicationController
   before_action :authenticate_user!
   authorize_resource
-  before_action :set_product, only: [:show, :edit, :update, :destroy, :create_variants]
+  before_action :set_product, only: [:show, :edit, :update, :destroy, :create_variants, :image_upload]
 
   # GET /products
   def index
@@ -13,12 +13,13 @@ class ProductsController < ApplicationController
 
   # GET /products/1
   def show
+    redirect_to products_url, notice: "Не туда пошёл"
   end
 
   # GET /products/new
   def new
     @product = Product.new
-    @product.variants.build
+    # @product.variants.build
   end
 
   # GET /products/1/edit
@@ -46,7 +47,7 @@ class ProductsController < ApplicationController
   def update
     respond_to do |format|
       if @product.update(product_params)
-        format.html { redirect_to @product, notice: "Product was successfully updated." }
+        format.html { redirect_to products_url, notice: "Product was successfully updated." }
         format.json { render :show, status: :ok, location: @product }
       else
         format.html { render :edit, status: :unprocessable_entity }
@@ -76,32 +77,93 @@ class ProductsController < ApplicationController
     end
   end
 
-  def delete_image
-    ActiveStorage::Attachment.where(id: params[:image_id])[0].purge
+  def image_upload
+    params.require(:blob_signed_id)
+    signed_id = params["blob_signed_id"]
+    upload_blob = ActiveStorage::Blob.find_signed(signed_id)
+    filename = upload_blob.filename
+    content_type = upload_blob.content_type
+    @product.images.attach(upload_blob)
+
+    @blob = upload_blob
+
     respond_to do |format|
-      #format.html { redirect_to edit_product_path(params[:id]), notice: 'Image was successfully deleted.' }
-      format.json { render json: { :status => "ok", :message => "destroyed" } }
+      format.turbo_stream { flash.now[:notice] = t(".success") }
+    end
+  end
+  
+  def delete_image
+    @blob_id = params[:blob_id]
+
+    ActiveStorage::Attachment.find(params[:blob_id]).destroy
+
+    respond_to do |format|
+      flash.now[:notice] = t(".success")
+      format.turbo_stream  do
+        render turbo_stream: [
+          turbo_stream.remove(@blob_id),
+          render_turbo_flash
+        ]
+      end
     end
   end
 
   def import
-    Rails.env.development? ? Services::Import.product : ImportProductJob.perform_later
-    redirect_to products_path, notice: 'Запущен процесс Обновление Товаров InSales. Дождитесь выполнении'
+    ImportProductJob.perform_later
+    respond_to do |format|
+      flash.now[:notice] = "Запущен процесс Обновление Товаров Дождитесь выполнении"
+      format.turbo_stream  do
+        render turbo_stream: [
+          render_turbo_flash
+        ]
+      end
+    end
   end
 
   def avito
-    Rails.env.development? ? Services::Export.avito : ExportAvitoJob.perform_later
-    redirect_to products_path, notice: 'Запущен процесс создания файла авито. Дождитесь выполнении'
+    ExportAvitoJob.perform_later
+    respond_to do |format|
+    flash.now[:notice] = "Запущен процесс создания файла авито. Дождитесь выполнении"
+    format.turbo_stream  do
+      render turbo_stream: [
+        render_turbo_flash
+      ]
+    end
+    end
   end
 
   def drom
-    Rails.env.development? ? Services::Export.drom : ExportDromJob.perform_later
-    redirect_to products_path, notice: 'Запущен процесс создания файла авито. Дождитесь выполнении'
+    ExportDromJob.perform_later
+    respond_to do |format|
+    flash.now[:notice] = "Запущен процесс создания файла drom. Дождитесь выполнении"
+    format.turbo_stream  do
+      render turbo_stream: [
+        render_turbo_flash
+      ]
+    end
+    end
   end
 
   def create_variants
-    Rails.env.development? ? Services::Import.create_variants(@product) : ProductVariantJob.perform_later(@product)
-    redirect_to @product, notice: 'Запущен процесс создания вариантов. Дождитесь выполнении'
+    # ProductVariantJob.perform_later(@product.id)
+    @var_count = 10 #29
+    if @product.variants.size < @var_count
+        start_count = @product.variants.size
+        calc_loop = @var_count-start_count
+        for a in 1..calc_loop do
+            variant = @product.variants.create!(status: "New", sku: @product.sku, title: @product.title, desc: @product.desc, period: start_count+a)
+            variant.update!(status: "Process")
+            VariantImageJob.perform_later(@product.id, variant.id) if @product.images.present?
+        end
+    end
+    respond_to do |format|
+      flash.now[:notice] = "Запустили"
+      format.turbo_stream  do
+        render turbo_stream: [
+          render_turbo_flash
+        ]
+      end
+    end
   end
 
   private
@@ -112,6 +174,6 @@ class ProductsController < ApplicationController
 
     # Only allow a trusted parameter "white list" through.
     def product_params
-      params.require(:product).permit(:sku, :title, :desc, :quantity, :costprice, :price, :offer_id,:barcode, :avito_param, :avito_date_begin, images: [], images_attachments_attributes: [:id, :_destroy], :variants_attributes =>[:id, :sku, :product_id, :title, :desc, :period, :_destroy])
+      params.require(:product).permit(:sku, :title, :desc, :quantity, :costprice, :price, :offer_id,:barcode, :avito_param, :avito_date_begin, images: [], :variants_attributes =>[:id, :status, :sku, :product_id, :title, :desc, :period, :_destroy])
     end
 end
