@@ -43,14 +43,14 @@ Rails.application.configure do
   config.assets.debug = true
 
   config.action_mailer.smtp_settings = {
-    address: Rails.application.secrets.email_provider_address,
-    port: Rails.application.secrets.email_provider_port,
-    domain: Rails.application.secrets.domain_name,
+    address: "email_provider_address",
+    port: "email_provider_port",
+    domain: "domain_name",
     authentication: "plain",
     tls: true,
     enable_starttls_auto: true,
-    user_name: Rails.application.secrets.email_provider_username,
-    password: Rails.application.secrets.email_provider_password
+    user_name: "email_provider_username",
+    password: "email_provider_password"
   }
   # ActionMailer Config
   config.action_mailer.default_url_options = { :host => 'localhost:3000' }
@@ -68,9 +68,9 @@ Rails.application.configure do
 
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
-  config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+  # config.file_watcher = ActiveSupport::EventedFileUpdateChecker
 
   config.active_storage.service = :local
+
 end
 
-Rails.application.routes.default_url_options[:host] = 'localhost:3000'
