@@ -1,4 +1,11 @@
 Rails.application.routes.draw do
+
+  require 'sidekiq/web'
+
+  authenticate :user, ->(user) { user.admin? } do
+    mount Sidekiq::Web => '/sidekiq'
+  end
+
   resources :variants do
     collection do
       post '/:id/create_images', action: 'create_images', as: 'create_images'
@@ -12,7 +19,10 @@ Rails.application.routes.draw do
       get :avito
       get :drom
       post '/:id/create_variants', action: 'create_variants', as: 'create_variants'
-      delete '/:id/images/:image_id', action: 'delete_image', as: 'delete_image'
+      # delete '/:id/images/:image_id', action: 'delete_image', as: 'delete_image'
+      post '/:id/image_upload', action: 'image_upload', as: 'image_upload'
+      # post :image_upload
+      post :delete_image
     end
   end
 
@@ -37,9 +47,5 @@ Rails.application.routes.draw do
       delete '/:id/images/:image_id', action: 'delete_image', as: 'delete_image'
     end
   end
-
-  authenticated :user, -> user { user.admin? }  do
-    mount DelayedJobWeb, at: "/job"
-  end
-
+  
 end
